requirements.txt
apify-client==1.8.0
requests==2.31.0
